get user/myRooms and chat/allRooms - formatted
