+++ backend/src/user/user.controller.ts
@@ -104,4 +104,10 @@ export class UserController {
   async handleGetUserByEmail(@Query() dto: GetUserByEmailDTO) {
     return await this.userService.getUserByEmail(dto);
   }
+
+  @Get('myRooms')
+  @ApiOkResponse({ description: 'Returns rooms that user is connected to' })
+  async handleGetRooms(@GetUser() user: User) {
+    return this.userService.getRooms(user);
+  }
 }
